Pass all tests

+++ test/supply_chain.test.js
@@ -93,7 +93,7 @@ contract("SupplyChain", function (accounts) {
           "Struct Item should have a `sku` member"
         );
         assert(
-          isType(subjectStruct)("sku")("uint"), 
+          isType(subjectStruct)("sku")("uint256"), 
           "`sku` should be of type `uint`"
         );
       });
@@ -104,7 +104,7 @@ contract("SupplyChain", function (accounts) {
           "Struct Item should have a `price` member"
         );
         assert(
-          isType(subjectStruct)("price")("uint"), 
+          isType(subjectStruct)("price")("uint256"), 
           "`price` should be of type `uint`"
         );
       });
